[64] Fix ClientProfileScreen eslint errors and create SidebarComponent
