+++ src/Components/GenericRegisterScreen/style.js
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import { IoPersonCircleOutline } from 'react-icons/io5';
 import colors from '../../Constants/colors';
 
 export const Main = styled.div` 
@@ -35,26 +34,6 @@ export const Container = styled.div`
         flex-direction: column;
         justify-content: flex-start;
         padding-top: 10vh;
-    }
-`;
-
-export const Sidebar = styled.div`
-
-    width: 30%;
-    height: 100%;
-    background: ${colors.primary};
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    overflow: hidden;
-
-    @media(max-width: 720px){
-        width: 100%;
-        height: 20%;
-        background-color: ${colors.secondary};
-        align-content: center;
-        justify-content: center;
-   
     }
 `;
 
@@ -115,41 +94,6 @@ export const DivButtom = styled.div`
         margin-right: 0%;
 
         
-    }
-`;
-
-export const SidebarText = styled.div`
-
-    color: ${colors.secondary};
-    display: flex;
-    flex-wrap: wrap;
-    flex-direction: column;
-    align-items: center;
-
-    @media(max-width: 720px){
-        visibility: hidden;
-        height: 0%;
-    }
-
-    p {
-        font: Open Arial;
-        font-size: 1.5vw;
-
-    }
-`;
-
-export const Icon = styled(IoPersonCircleOutline)`
-
-    width: 100%;
-    height: 40%;
-    color: ${colors.secondary};
-    margin-top: 10px;
-
-    @media(max-width: 720px){
-        width: 100%;
-        height: 100%;
-
-        color: ${colors.primary};
     }
 `;
